test: consolidate entitlement coverage
